chat_bot/pyproject.toml
[tool.poetry]
name = "chat-bot"
version = "0.1.0"
description = ""
authors = ["seiya <ilya_shtak@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.13"
g4f = "^0.3.9.7"
python-telegram-bot = "^21.9"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
